+++ Backend/src/controller/userController.js
@@ -3,12 +3,16 @@ const UserService = require("../service/userService");
 const {headersCors} = require("../headersCors");
 const Role = require("../model/role");
 const UserClass = require("../model/user");
+const {validateRightIsWrong, validateRightIsWrongCheckRole} = require("../middleware/authentication");
 
 module.exports = function register() {
     console.log("user controller pages loaded")
 }
 
 Router.get("user", async function getUser(req, res) {
+    if(validateRightIsWrong(req, res)){
+        return
+    }
     const body = req.body.username;
     const response = await UserService.getUserByName(body)
     res.writeHead(200, headersCors)
@@ -16,6 +20,9 @@ Router.get("user", async function getUser(req, res) {
 })
 
 Router.get("users", async function getUser(req, res) {
+    if(validateRightIsWrong(req, res)){
+        return
+    }
     const response = await UserService.getUsers()
     res.writeHead(200, headersCors)
     res.end(JSON.stringify(response.map(r => {
@@ -24,6 +31,9 @@ Router.get("users", async function getUser(req, res) {
 })
 
 Router.post("user", async function addUser(req, res) {
+    if(validateRightIsWrongCheckRole(req, res)){
+        return
+    }
     const body = req.body;
     const found = await UserService.getUserByName(body.username)
     if (found.length <= 0) {
@@ -38,6 +48,9 @@ Router.post("user", async function addUser(req, res) {
 })
 
 Router.delete("user", async function deleteUser(req, res) {
+    if(validateRightIsWrong(req, res)){
+        return
+    }
     const body = req.body;
     const response = await UserService.deleteUserById(body)
     res.writeHead(200, headersCors)
@@ -45,6 +58,9 @@ Router.delete("user", async function deleteUser(req, res) {
 })
 
 Router.post("user/update", async function update(req, res) {
+    if(validateRightIsWrong(req, res)){
+        return
+    }
     const body = req.body;
     const response = await UserService.updateUser(body.id, body)
     res.writeHead(200, headersCors)
